Clean up duplicated planner logic for window operator
